Split aggregator_api/src/lib.rs
